Initial Revision

+++ app/src/main/java/au/com/mazeit/seedaudit/SeedAudit.java
@@ -7,9 +7,12 @@ package au.com.mazeit.seedaudit;
 public class SeedAudit {
     public int slnId;
     public int locationId;
-    public float gross;
-    public float barrelWgt;
-    public float nett;
+    //public float gross;
+    //public float barrelWgt;
+    //public float nett;
+    public double gross;
+    public double barrelWgt;
+    public double nett;
     public boolean verified;
     public String changedBy;
     public String ChangedDate;
